# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Hubuum REST API",
		"description": "OpenAPI documentation for the Hubuum REST service.",
		"license": {
			"name": "MIT",
			"identifier": "MIT"
		},
		"version": "0.0.1"
	},
	"servers": [
		{
			"url": "/",
			"description": "Current deployment base URL"
		}
	],
	"paths": {
		"/api/v0/auth/login": {
			"post": {
				"tags": ["auth"],
				"summary": "Post Api V0 Auth Login",
				"description": "Auto-generated documentation for POST /api/v0/auth/login.",
				"operationId": "postApiV0AuthLogin",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginUser"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Token issued",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"429": {
						"description": "Too many login attempts",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"ApiErrorResponse": {
				"type": "object",
				"required": ["error", "message"],
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"error": "Unauthorized",
					"message": "Authentication failure"
				}
			},
...
			"LoginResponse": {
				"type": "object",
				"required": ["token"],
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"example": {
					"token": "eyJhbGciOi...example-token"
				}
			},
			"LoginUser": {
				"type": "object",
				"description": "Struct to log in a user.\n\nThe password is expected to be plaintext.",
				"required": ["username", "password"],
				"properties": {
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"example": {
					"password": "correct-horse-battery-staple",
					"username": "alice"
				}
			},
...
		},
		"securitySchemes": {
			"bearer_auth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
...
}